requirements.txt
pandas==2.2.2
numpy==1.26.4
streamlit ==1.35.0
pip==24.0
joblib
pyreadstat
pillow
streamlit-option-menu
scikit-learn==1.5.0.